txBlock spike
