+++ chain/tx_block.go
@@ -5,15 +5,11 @@ package chain
 
 import (
 	"context"
-	"encoding/binary"
-	"errors"
 	"fmt"
 	"time"
 
 	"github.com/ava-labs/avalanchego/ids"
-	"github.com/ava-labs/avalanchego/snow/validators"
 	"github.com/ava-labs/avalanchego/utils/set"
-	"github.com/ava-labs/avalanchego/vms/platformvm/warp"
 	"github.com/ava-labs/avalanchego/x/merkledb"
 	"go.opentelemetry.io/otel/attribute"
 	oteltrace "go.opentelemetry.io/otel/trace"
@@ -21,28 +17,15 @@ import (
 
 	"github.com/ava-labs/hypersdk/codec"
 	"github.com/ava-labs/hypersdk/consts"
-	"github.com/ava-labs/hypersdk/tstate"
 	"github.com/ava-labs/hypersdk/utils"
-	"github.com/ava-labs/hypersdk/window"
 	"github.com/ava-labs/hypersdk/workers"
 )
 
 type TxBlock struct {
-	Prnt   ids.ID `json:"parent"`
-	Tmstmp int64  `json:"timestamp"`
-	Hght   uint64 `json:"height"`
-
-	// TODO: need to put in state (because won't know when parsing block)
-	// UnitPrice  uint64        `json:"unitPrice"`
-	// UnitWindow window.Window `json:"unitWindow"`
-
-	// PChainHeight is needed to verify warp messages
-	ContainsWarp bool   `json:"containsWarp"`
-	PChainHeight uint64 `json:"pchainHeight"`
-
-	Txs []*Transaction `json:"txs"`
-	// TODO: store in state
-	// WarpResults set.Bits64     `json:"warpResults"`
+	Prnt   ids.ID         `json:"parent"`
+	Tmstmp int64          `json:"timestamp"`
+	Hght   uint64         `json:"height"`
+	Txs    []*Transaction `json:"txs"`
 
 	// TEMP
 	Issued int64 `json:"issued"`
@@ -54,19 +37,13 @@ type TxBlock struct {
 type StatelessTxBlock struct {
 	*TxBlock `json:"block"`
 
-	id     ids.ID
-	t      time.Time
-	bytes  []byte
+	id    ids.ID
+	t     time.Time
+	bytes []byte
+
 	txsSet set.Set[ids.ID]
 
-	warpMessages map[ids.ID]*warpJob
-	vdrState     validators.State
-
-	results []*Result
-
-	vm        VM
-	state     merkledb.TrieView
-	processor *Processor
+	vm VM
 
 	sigJob *workers.Job
 }
@@ -86,16 +63,6 @@ func NewTxBlock(ectx *TxExecutionContext, vm VM, parent *StatelessTxBlock, tmstm
 	}
 }
 
-// warpJob is used to signal to a listner that a *warp.Message has been
-// verified.
-type warpJob struct {
-	msg          *warp.Message
-	signers      int
-	verifiedChan chan bool
-	verified     bool
-	warpNum      int
-}
-
 // populateTxs is only called on blocks we did not build
 func (b *StatelessTxBlock) populateTxs(ctx context.Context) error {
 	ctx, span := b.vm.Tracer().Start(ctx, "StatelessBlock.populateTxs")
@@ -103,7 +70,8 @@ func (b *StatelessTxBlock) populateTxs(ctx context.Context) error {
 
 	// Setup signature verification job
 	var sspan oteltrace.Span
-	if b.vm.GetVerifySignatures() {
+	verifySignatures := b.vm.GetVerifySignatures()
+	if verifySignatures {
 		job, err := b.vm.Workers().NewJob(len(b.Txs))
 		if err != nil {
 			return err
@@ -114,45 +82,16 @@ func (b *StatelessTxBlock) populateTxs(ctx context.Context) error {
 
 	// Process transactions
 	b.txsSet = set.NewSet[ids.ID](len(b.Txs))
-	b.warpMessages = map[ids.ID]*warpJob{}
 	for _, tx := range b.Txs {
-		if b.vm.GetVerifySignatures() {
+		if verifySignatures {
 			b.sigJob.Go(tx.AuthAsyncVerify())
 		}
 		if b.txsSet.Contains(tx.ID()) {
 			return ErrDuplicateTx
 		}
 		b.txsSet.Add(tx.ID())
-
-		// Check if we need the block context to verify the block (which contains
-		// an Avalanche Warp Message)
-		//
-		// Instead of erroring out if a warp message is invalid, we mark the
-		// verification as skipped and include it in the verification result so
-		// that a fee can still be deducted.
-		if tx.WarpMessage != nil {
-			if !b.ContainsWarp {
-				return ErrWarpResultMismatch
-			}
-			if len(b.warpMessages) == MaxWarpMessages {
-				return ErrTooManyWarpMessages
-			}
-			signers, err := tx.WarpMessage.Signature.NumSigners()
-			if err != nil {
-				return err
-			}
-			b.warpMessages[tx.ID()] = &warpJob{
-				msg:          tx.WarpMessage,
-				signers:      signers,
-				verifiedChan: make(chan bool, 1),
-				warpNum:      len(b.warpMessages),
-			}
-		}
 	}
-	if len(b.warpMessages) > 0 && !b.ContainsWarp {
-		return ErrWarpResultMismatch
-	}
-	if b.vm.GetVerifySignatures() {
+	if verifySignatures {
 		b.sigJob.Done(func() { sspan.End() })
 	}
 	return nil
@@ -174,11 +113,6 @@ func ParseTxBlock(
 		}
 		if len(blk.Txs) == 0 {
 			return nil, ErrNoTxs
-		}
-		r := vm.Rules(blk.Tmstmp)
-		// TODO: transition to block size
-		if blk.UnitsConsumed > r.GetMaxTxBlockUnits() {
-			return nil, ErrBlockTooBig
 		}
 	}
 
@@ -226,8 +160,6 @@ func (b *StatelessTxBlock) initializeBuilt(
 	b.bytes = blk
 	b.id = utils.ToID(b.bytes)
 	b.t = time.Unix(b.TxBlock.Tmstmp, 0)
-	b.state = state
-	b.results = results
 	b.txsSet = set.NewSet[ids.ID](len(b.Txs))
 	for _, tx := range b.Txs {
 		b.txsSet.Add(tx.ID())
@@ -238,56 +170,12 @@ func (b *StatelessTxBlock) initializeBuilt(
 // implements "snowman.Block.choices.Decidable"
 func (b *StatelessTxBlock) ID() ids.ID { return b.id }
 
-func preVerifyWarpMessage(msg *warp.Message, chainID ids.ID, r Rules) (uint64, uint64, error) {
-	if msg.DestinationChainID != chainID && msg.DestinationChainID != ids.Empty {
-		return 0, 0, ErrInvalidChainID
-	}
-	if msg.SourceChainID == chainID {
-		return 0, 0, ErrInvalidChainID
-	}
-	if msg.SourceChainID == msg.DestinationChainID {
-		return 0, 0, ErrInvalidChainID
-	}
-	allowed, num, denom := r.GetWarpConfig(msg.SourceChainID)
-	if !allowed {
-		return 0, 0, ErrDisabledChainID
-	}
-	return num, denom, nil
-}
-
-// verifyWarpMessage will attempt to verify a given warp message provided by an
-// Action.
-func (b *StatelessTxBlock) verifyWarpMessage(ctx context.Context, r Rules, msg *warp.Message) bool {
-	warpID := utils.ToID(msg.Payload)
-	num, denom, err := preVerifyWarpMessage(msg, b.vm.ChainID(), r)
-	if err != nil {
-		b.vm.Logger().
-			Warn("unable to verify warp message", zap.Stringer("warpID", warpID), zap.Error(err))
-		return false
-	}
-	if err := msg.Signature.Verify(
-		ctx,
-		&msg.UnsignedMessage,
-		b.vdrState,
-		b.PChainHeight,
-		num,
-		denom,
-	); err != nil {
-		b.vm.Logger().
-			Warn("unable to verify warp message", zap.Stringer("warpID", warpID), zap.Error(err))
-		return false
-	}
-	return true
-}
-
 func (b *StatelessTxBlock) Verify(ctx context.Context) error {
 	ctx, span := b.vm.Tracer().Start(
 		ctx, "StatelessTxBlock.Verify",
 		oteltrace.WithAttributes(
 			attribute.Int("txs", len(b.Txs)),
 			attribute.Int64("height", int64(b.Hght)),
-			attribute.Bool("containsWarp", b.ContainsWarp),
-			attribute.Int64("pchainHeight", int64(b.PChainHeight)),
 		),
 	)
 	defer span.End()
@@ -304,8 +192,6 @@ func (b *StatelessTxBlock) Verify(ctx context.Context) error {
 		return ErrTimestampTooLate
 	case len(b.Txs) == 0:
 		return ErrNoTxs
-	case b.UnitsConsumed > r.GetMaxTxBlockUnits():
-		return ErrBlockTooBig
 	}
 
 	// Verify parent is verified and available
@@ -321,22 +207,6 @@ func (b *StatelessTxBlock) Verify(ctx context.Context) error {
 	if b.Tmstmp < parent.Tmstmp {
 		return ErrTimestampTooEarly
 	}
-	ectx, err := GenerateTxExecutionContext(ctx, b.vm.ChainID(), b.Tmstmp, parent, b.vm.Tracer(), r)
-	if err != nil {
-		return err
-	}
-	switch {
-	case b.UnitPrice != ectx.NextUnitPrice:
-		return ErrInvalidUnitPrice
-	case b.UnitWindow != ectx.NextUnitWindow:
-		b.vm.Logger().Warn("unit window mismatch", zap.Uint64("height", b.Hght), zap.Stringer("parent", parent.ID()), zap.Binary("found", b.UnitWindow[:]), zap.Binary("expected", ectx.NextUnitWindow[:]))
-		return ErrInvalidUnitWindow
-	}
-	log.Info(
-		"verify context",
-		zap.Uint64("height", b.Hght),
-		zap.Uint64("unit price", b.UnitPrice),
-	)
 
 	// Ensure tx cannot be replayed
 	//
@@ -355,124 +225,8 @@ func (b *StatelessTxBlock) Verify(ctx context.Context) error {
 		return fmt.Errorf("%w: duplicate in ancestry", ErrDuplicateTx)
 	}
 
-	// Start validating warp messages, if they exist
-	var invalidWarpResult bool
-	if b.ContainsWarp {
-		_, sspan := b.vm.Tracer().Start(ctx, "StatelessTxBlock.verifyWarpMessages")
-		b.vdrState = b.vm.ValidatorState()
-		go func() {
-			defer sspan.End()
-			// We don't use [b.vm.Workers] here because we need the warp verification
-			// results during normal execution. If we added a job to the workers queue,
-			// it would get executed after all signatures. Additionally, BLS
-			// Multi-Signature verification is already parallelized so we should just
-			// do one at a time to avoid overwhelming the CPU.
-			for txID, msg := range b.warpMessages {
-				if ctx.Err() != nil {
-					return
-				}
-				blockVerified := b.WarpResults.Contains(uint(msg.warpNum))
-				if b.vm.IsBootstrapped() && !invalidWarpResult {
-					start := time.Now()
-					verified := b.verifyWarpMessage(ctx, r, msg.msg)
-					msg.verifiedChan <- verified
-					msg.verified = verified
-					log.Info(
-						"processed warp message",
-						zap.Stringer("txID", txID),
-						zap.Bool("verified", verified),
-						zap.Int("signers", msg.signers),
-						zap.Duration("t", time.Since(start)),
-					)
-					if blockVerified != verified {
-						invalidWarpResult = true
-					}
-				} else {
-					// When we are bootstrapping, we just use the result in the block.
-					//
-					// We also use the result in the block when we have found
-					// a verification mismatch (our verify result is different than the
-					// block) to avoid doing extra work.
-					msg.verifiedChan <- blockVerified
-					msg.verified = blockVerified
-				}
-			}
-		}()
-	}
-
-	// Optimisticaly fetch state
-	if parent.Last {
-		state, err := parent.ChildState(ctx, 30_000)
-		if err != nil {
-			return err
-		}
-		b.processor = NewProcessor(b.vm.Tracer(), state, tstate.New(30_000))
-	} else {
-		b.processor = parent.processor
-	}
-	b.processor.Prefetch(ctx, b)
-
-	// Process new transactions
-	unitsConsumed, results, stateChanges, stateOps, err := b.processor.Execute(ctx, ectx, r)
-	if err != nil {
-		log.Error("failed to execute block", zap.Error(err))
-		return err
-	}
-	b.vm.RecordStateChanges(stateChanges)
-	b.vm.RecordStateOperations(stateOps)
-	if b.UnitsConsumed != unitsConsumed {
-		return fmt.Errorf(
-			"%w: required=%d found=%d",
-			ErrInvalidUnitsConsumed,
-			unitsConsumed,
-			b.UnitsConsumed,
-		)
-	}
-
-	// Ensure warp results are correct
-	if invalidWarpResult {
-		return ErrWarpResultMismatch
-	}
-	numWarp := len(b.warpMessages)
-	if numWarp > MaxWarpMessages {
-		return ErrTooManyWarpMessages
-	}
-	var warpResultsLimit set.Bits64
-	warpResultsLimit.Add(uint(numWarp))
-	if b.WarpResults >= warpResultsLimit {
-		// If the value of [WarpResults] is greater than the value of uint64 with
-		// a 1-bit shifted [numWarp] times, then there are unused bits set to
-		// 1 (which should is not allowed).
-		return ErrWarpResultMismatch
-	}
-
 	// We purposely avoid root calc (which would make this spikey)
 	b.vm.RecordTxBlockVerify(time.Since(start))
-
-	// Compute state root if last
-	//
-	// TODO: better protect against malicious usage of this (may need to be
-	// invoked by caller block)
-	if b.Last {
-		// Commit to base
-		if err := b.processor.Commit(ctx); err != nil {
-			return err
-		}
-		// Store height in state to prevent duplicate roots
-		base := b.processor.db
-		if err := base.Insert(ctx, b.vm.StateManager().HeightKey(), binary.BigEndian.AppendUint64(nil, b.Hght)); err != nil {
-			return err
-		}
-		start := time.Now()
-		if _, err := base.GetMerkleRoot(ctx); err != nil {
-			return err
-		}
-		b.vm.RecordRootCalculated(time.Since(start))
-		b.state = base
-	}
-
-	// We wait for signatures in root block.
-	b.results = results
 	return nil
 }
 
@@ -488,6 +242,7 @@ func (b *StatelessTxBlock) Accept(ctx context.Context) error {
 func (b *StatelessTxBlock) Reject(ctx context.Context) error {
 	ctx, span := b.vm.Tracer().Start(ctx, "StatelessTxBlock.Reject")
 	defer span.End()
+
 	return nil
 }
 
@@ -548,63 +303,20 @@ func (b *StatelessTxBlock) GetTimestamp() int64 {
 	return b.Tmstmp
 }
 
-func (b *StatelessTxBlock) GetUnitPrice() uint64 {
-	return b.UnitPrice
-}
-
-func (b *StatelessTxBlock) Results() []*Result {
-	return b.results
-}
-
-// We assume this will only be called once we are done syncing, so it is safe
-// to assume we will eventually get to a block with state.
-func (b *StatelessTxBlock) ChildState(
-	ctx context.Context,
-	estimatedChanges int,
-) (merkledb.TrieView, error) {
-	ctx, span := b.vm.Tracer().Start(ctx, "StatelessTxBlock.childState")
-	defer span.End()
-
-	// Return committed state if block is accepted or this is genesis.
-	if b.Hght <= b.vm.LastAcceptedBlock().MaxTxHght() {
-		state, err := b.vm.State()
-		if err != nil {
-			return nil, err
-		}
-		return state.NewPreallocatedView(estimatedChanges)
-	}
-
-	// Process block if not yet processed and not yet accepted.
-	//
-	// We don't need to handle the case where the tx block is loaded from disk
-	// because that will hit the first if check here.
-	if b.state == nil {
-		return nil, errors.New("not implemented")
-	}
-	return b.state.NewPreallocatedView(estimatedChanges)
-}
-
 func (b *TxBlock) Marshal(
 	actionRegistry ActionRegistry,
 	authRegistry AuthRegistry,
 ) ([]byte, error) {
-	size := consts.IDLen + consts.Uint64Len + consts.Uint64Len + consts.Uint64Len +
-		window.WindowSliceSize + codec.BoolLen + consts.Uint64Len + consts.IntLen
+	size := consts.IDLen + consts.Uint64Len + consts.Uint64Len + consts.IntLen
 	for _, tx := range b.Txs {
 		size += tx.Size()
 	}
-	size += consts.Uint64Len + consts.Uint64Len + codec.BoolLen + consts.Uint64Len
+	size += consts.Uint64Len
 	p := codec.NewWriter(size, consts.NetworkSizeLimit)
 
 	p.PackID(b.Prnt)
 	p.PackInt64(b.Tmstmp)
 	p.PackUint64(b.Hght)
-
-	p.PackUint64(b.UnitPrice)
-	p.PackWindow(b.UnitWindow)
-
-	p.PackBool(b.ContainsWarp)
-	p.PackUint64(b.PChainHeight)
 
 	p.PackInt(len(b.Txs))
 	for _, tx := range b.Txs {
@@ -612,11 +324,8 @@ func (b *TxBlock) Marshal(
 			return nil, err
 		}
 	}
-	p.PackUint64(uint64(b.WarpResults))
-	p.PackUint64(b.UnitsConsumed)
-	p.PackBool(b.Last)
-	p.PackInt64(b.Issued)
 
+	p.PackInt64(b.Issued)
 	return p.Bytes(), p.Err()
 }
 
@@ -630,16 +339,6 @@ func UnmarshalTxBlock(raw []byte, parser Parser) (*TxBlock, error) {
 	b.Tmstmp = p.UnpackInt64(false)
 	b.Hght = p.UnpackUint64(false)
 
-	b.UnitPrice = p.UnpackUint64(false)
-	p.UnpackWindow(&b.UnitWindow)
-
-	b.ContainsWarp = p.UnpackBool()
-	b.PChainHeight = p.UnpackUint64(false)
-	if err := p.Err(); err != nil {
-		// Check that header was parsed properly before unwrapping transactions
-		return nil, err
-	}
-
 	// Parse transactions
 	txCount := p.UnpackInt(false) // could be 0 in genesis
 	actionRegistry, authRegistry := parser.Registry()
@@ -651,11 +350,8 @@ func UnmarshalTxBlock(raw []byte, parser Parser) (*TxBlock, error) {
 		}
 		b.Txs = append(b.Txs, tx)
 	}
-	b.WarpResults = set.Bits64(p.UnpackUint64(false))
-	b.UnitsConsumed = p.UnpackUint64(false)
-	b.Last = p.UnpackBool()
-	b.Issued = p.UnpackInt64(false)
 
+	b.Issued = p.UnpackInt64(false)
 	if !p.Empty() {
 		// Ensure no leftover bytes
 		return nil, ErrInvalidObject
